+++ Sources/Sentry/PrivateSentrySDKOnly.m
@@ -282,7 +282,12 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
 + (NSArray<NSData *> *)captureScreenshots
 {
 #if SENTRY_TARGET_REPLAY_SUPPORTED
-    return [SentryDependencyContainer.sharedInstance.screenshot appScreenshotsData];
+    // As the options are not passed in by the hybrid SDK, we need to use the options from the
+    // current hub.
+    SentryViewScreenshotOptions *_Nonnull options = PrivateSentrySDKOnly.options.screenshot;
+    SentryScreenshotSource *_Nonnull screenshotSource =
+        [SentryDependencyContainer.sharedInstance getScreenshotSourceForOptions:options];
+    return [screenshotSource appScreenshotsData];
 #else
     SENTRY_LOG_DEBUG(
         @"PrivateSentrySDKOnly.captureScreenshots only works with UIKit enabled. Ensure you're "